Only removed another warning
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@129 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/filter.c
@@ -119,7 +119,7 @@ filter* filter_lp_fir_alloc(double freq, int order) {
 	/* adjust the gain to be exact 1.0 */
 	for(i=0;i<=midorder;++i) {
 #ifdef FILTER_USE_INT
-		f->xcoeffs[i] /= ((int)(gain+.5) != 0) ? (int)(gain+.5) : gain;
+		f->xcoeffs[i] /= ((int)(gain+.5) != 0) ? (int)(gain+.5) : (int) gain;
 #else
 		f->xcoeffs[i] = f->xcoeffs[i] * (double)(1 << FILTER_INT_FRAC) / gain;
 #endif
